refactor: updateModalData 함수 위치 변경 및 renderModal.ts로 이동
